+++ account_invoice_triple_discount/i18n/es_EC.po
@@ -1,0 +1,39 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# * account_invoice_triple_discount
+# 
+# Translators:
+# OCA Transbot <transbot@odoo-community.org>, 2017
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 10.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2017-08-09 02:45+0000\n"
+"PO-Revision-Date: 2017-08-09 02:45+0000\n"
+"Last-Translator: OCA Transbot <transbot@odoo-community.org>, 2017\n"
+"Language-Team: Spanish (Ecuador) (https://www.transifex.com/oca/teams/23907/es_EC/)\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Language: es_EC\n"
+"Plural-Forms: nplurals=2; plural=(n != 1);\n"
+
+#. module: account_invoice_triple_discount
+#: model:ir.model.fields,field_description:account_invoice_triple_discount.field_account_invoice_line_discount2
+msgid "Discount 2 (%)"
+msgstr ""
+
+#. module: account_invoice_triple_discount
+#: model:ir.model.fields,field_description:account_invoice_triple_discount.field_account_invoice_line_discount3
+msgid "Discount 3 (%)"
+msgstr ""
+
+#. module: account_invoice_triple_discount
+#: model:ir.model,name:account_invoice_triple_discount.model_account_invoice
+msgid "Invoice"
+msgstr "Factura"
+
+#. module: account_invoice_triple_discount
+#: model:ir.model,name:account_invoice_triple_discount.model_account_invoice_line
+msgid "Invoice Line"
+msgstr "Detalle de Factura"
