+++ account_invoice_triple_discount/models/account_invoice.py
@@ -1,25 +1,64 @@
 # -*- coding: utf-8 -*-
 # Copyright 2017 Tecnativa - David Vidal
+# Copyright 2017 Tecnativa - Pedro M. Baeza
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from odoo import models
+from odoo import api, fields, models
+from odoo.addons import decimal_precision as dp
 
 
 class AccountInvoice(models.Model):
     _inherit = "account.invoice"
 
     def get_taxes_values(self):
-        prices = {}
-        discount = {}
+        vals = {}
         for line in self.invoice_line_ids:
-            prices[line.id] = line.price_unit
-            discount[line.id] = line.discount
-            line.price_unit *= (1 - (line.discount or 0.0) / 100.0)
-            line.price_unit *= (1 - (line.discount2 or 0.0) / 100.0)
-            line.price_unit *= (1 - (line.discount3 or 0.0) / 100.0)
-            line['discount'] = 0.0
+            vals[line] = {
+                'price_unit': line.price_unit,
+                'discount': line.discount,
+            }
+            price_unit = line.price_unit * (1 - (line.discount or 0.0) / 100.0)
+            price_unit *= (1 - (line.discount2 or 0.0) / 100.0)
+            price_unit *= (1 - (line.discount3 or 0.0) / 100.0)
+            line.update({
+                'price_unit': price_unit,
+                'discount': 0.0,
+            })
         tax_grouped = super(AccountInvoice, self).get_taxes_values()
         for line in self.invoice_line_ids:
-            line['price_unit'] = prices[line.id]
-            line['discount'] = discount[line.id]
+            line.update(vals[line])
         return tax_grouped
+
+
+class AccountInvoiceLine(models.Model):
+    _inherit = "account.invoice.line"
+
+    discount2 = fields.Float(
+        'Discount 2 (%)',
+        digits=dp.get_precision('Discount'),
+        default=0.0,
+    )
+    discount3 = fields.Float(
+        'Discount 3 (%)',
+        digits=dp.get_precision('Discount'),
+        default=0.0,
+    )
+
+    @api.multi
+    @api.depends('discount2', 'discount3')
+    def _compute_price(self):
+        for line in self:
+            prev_price_unit = line.price_unit
+            prev_discount = line.discount
+            price_unit = line.price_unit * (1 - (line.discount or 0.0) / 100.0)
+            price_unit *= (1 - (line.discount2 or 0.0) / 100.0)
+            price_unit *= (1 - (line.discount3 or 0.0) / 100.0)
+            line.update({
+                'price_unit': price_unit,
+                'discount': 0.0,
+            })
+            super(AccountInvoiceLine, line)._compute_price()
+            line.update({
+                'price_unit': prev_price_unit,
+                'discount': prev_discount,
+            })
